update  迭代在线视频推理内容 实时查看

+++ wgai-module-system/wgai-system-biz/src/main/java/org/jeecg/modules/demo/video/util/identifyTypeNewOnnx.java
@@ -1409,7 +1409,8 @@ public class identifyTypeNewOnnx {
                     log.info("返回内容：" + ob);
                 }
 
-                if (pushInfo.getSaveRecord() != 0 && StringUtils.isNotEmpty(recordVideo)) {  //不保存本地录像
+                if (pushInfo.getSaveRecord() != 0 && StringUtils.isNotEmpty(recordVideo)&&pushInfo.getSaveLocalhost()!=0) {  //不保存本地录像
+                    log.info("[不保存本地录像]");
                     File imageFile = new File(recordVideo);
                     if (imageFile.exists()) {
                         imageFile.delete();
